moves store creation to dedicated file
